backend/requirements.txt
# NyayaSetu AI - Minimal Backend Requirements

# AI & LLM
google-generativeai>=0.4.0

# API Framework
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
python-multipart>=0.0.6
mangum>=0.17.0

# Utils
python-dotenv>=1.0.0
